Split `AnimationFrame` to it's own source files
