+++ NAS2D/Resource/AnimationFrame.h
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "../Duration.h"
+#include "../Math/Vector.h"
+#include "../Math/Rectangle.h"
+
+
+namespace NAS2D
+{
+	class Image;
+
+
+	struct AnimationFrame
+	{
+		const Image& image;
+		Rectangle<int> bounds;
+		Vector<int> anchorOffset;
+		Duration frameDelay;
+
+
+		bool isStopFrame() const;
+	};
+}
